# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Ubhi Backend API
  version: 1.0.0
  description: >
    API for the Ubhi brand site. Powers the public storefront (workshops, shop,
    gallery, journal, bookings, orders, newsletter and snail-mail signups) and
    the back-office admin dashboard (auth, users, customers, settings).
    Stateless JWT auth; role + per-section permission model. Storage is
    pluggable (zero-config JSON file store by default, PostgreSQL when
    DATABASE_URL is set).
  license:
    name: UNLICENSED
servers:
  - url: http://localhost:8090
    description: Local development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
    Ok:
      type: object
      properties:
        ok:
          type: boolean
          example: true
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: '#/components/schemas/AdminUser'
    ChangePasswordRequest:
      type: object
      required: [current, next]
      properties:
        current:
          type: string
        next:
          type: string
          minLength: 8
    AdminUser:
      type: object
      description: password_hash is never returned.
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        role:
          type: string
          enum: [owner, staff]
        permissions:
          type: array
          items:
            type: string
        name:
          type: string
        active:
          type: boolean
        created_at:
          type: string
          format: date-time
...
    CreateBookingRequest:
      type: object
      required: [name, email]
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        workshop_id:
          type: string
        workshop_title:
          type: string
        session_date:
          type: string
        price:
          type: number
        note:
          type: string
    Booking:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        workshop_id:
          type: string
          nullable: true
        workshop_title:
          type: string
        session_date:
          type: string
        price:
          type: number
        note:
          type: string
        status:
          type: string
        created_at:
          type: string
          format: date-time
...
    CreateCustomerRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        name:
          type: string
        phone:
          type: string
        address:
          type: string
        city:
          type: string
        postcode:
          type: string
        country:
          type: string
        marketing_opt_in:
          type: boolean
    UpdateCustomerRequest:
      type: object
      properties:
        email:
          type: string
        name:
          type: string
        phone:
          type: string
        address:
          type: string
        city:
          type: string
        postcode:
          type: string
        country:
          type: string
        marketing_opt_in:
          type: boolean
        active:
          type: boolean
        password:
          type: string
          minLength: 8
    Customer:
      type: object
      description: password_hash is never returned.
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        name:
          type: string
        phone:
          type: string
        address:
          type: string
        city:
          type: string
        postcode:
          type: string
        country:
          type: string
        marketing_opt_in:
          type: boolean
        active:
          type: boolean
        created_at:
          type: string
          format: date-time
    Workshop:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        blurb:
          type: string
        session_date:
          type: string
        price:
          type: number
        spots:
          type: integer
        image:
          type: string
        published:
          type: boolean
        created_at:
          type: string
          format: date-time
    WorkshopInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
        blurb:
          type: string
        session_date:
          type: string
        price:
          type: number
        spots:
          type: integer
        image:
          type: string
        published:
          type: boolean
...
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Missing or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Insufficient role or permission
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
paths:
...
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Log in (rate-limited 10 / 15 min per IP)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Token and user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          description: Too many login attempts
  /api/auth/me:
    get:
      tags: [Auth]
      summary: Current authenticated user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/AdminUser'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/auth/change-password:
    post:
      tags: [Auth]
      summary: Change own password
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ok'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/bookings:
    post:
      tags: [Bookings]
      summary: Create a booking (public reserve)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBookingRequest'
      responses:
        '201':
          description: Created booking
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '400':
          $ref: '#/components/responses/BadRequest'
    get:
      tags: [Bookings]
      summary: List bookings (section 'bookings')
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Booking'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/bookings/{id}:
    get:
      tags: [Bookings]
      summary: Get a booking (section 'bookings')
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Booking
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Bookings]
      summary: Update a booking (section 'bookings')
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Updated booking
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Bookings]
      summary: Delete a booking (section 'bookings')
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ok'
...
  /api/customers:
    post:
      tags: [Customers]
      summary: Register a storefront account (public)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCustomerRequest'
      responses:
        '201':
          description: Created customer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '400':
          $ref: '#/components/responses/BadRequest'
    get:
      tags: [Customers]
      summary: List customers (section 'customers')
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Customer'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/customers/{id}:
    patch:
      tags: [Customers]
      summary: Update a customer (section 'customers')
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCustomerRequest'
      responses:
        '200':
          description: Updated customer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Customers]
      summary: Delete a customer (section 'customers')
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ok'
  /api/workshops:
    get:
      tags: [Workshops]
      summary: List published workshops (public)
      responses:
        '200':
          description: Published workshops
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Workshop'
    post:
      tags: [Workshops]
      summary: Create a workshop (section 'workshops')
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkshopInput'
      responses:
        '201':
          description: Created workshop
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workshop'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /api/workshops/{id}:
    patch:
      tags: [Workshops]
      summary: Update a workshop (section 'workshops')
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkshopInput'
      responses:
        '200':
          description: Updated workshop
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workshop'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Workshops]
      summary: Delete a workshop (section 'workshops')
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ok'
...